split commoncrypto backend - hmac
